Consolidate htmldeps
